Revert typetype change
