+++ packages/hash/api/src/graphql/resolvers/entityType/entityTypeTypeFields.ts
@@ -3,10 +3,6 @@ import { Resolver, EntityType as GQLEntityType } from "../../apiTypes.gen";
 import { GraphQLContext } from "../../context";
 import { EntityType, UnresolvedGQLEntityType } from "../../../model";
 import { EntityTypeTypeFields } from "../../../db/adapter";
-import {
-  generateSchema$id,
-  schema$idRef,
-} from "../../../lib/schemas/jsonSchema";
 
 type EntityTypeMaybeTypeFields = UnresolvedGQLEntityType & {
   entityType?: GQLEntityType["entityType"];
@@ -23,27 +19,10 @@ const getEntityTypeType = async (dataSources: GraphQLContext["dataSources"]) =>
     throw new ApolloError(err.message);
   });
 
-const getEntityTypeById = async (
-  dataSources: GraphQLContext["dataSources"],
-  entityTypeId: string,
-) =>
-  EntityType.getEntityType(dataSources.db, { entityTypeId })
-    .then((type) => {
-      if (!type) {
-        throw new ApolloError(
-          `Could not find EntityType with entityTypeId ${entityTypeId}`,
-        );
-      }
-      return type.toGQLEntityType();
-    })
-    .catch((err) => {
-      throw new ApolloError(err.message);
-    });
-
 /**
  * Get the entityType of an EntityType, i.e. the "EntityType" EntityType
  */
-const entityTypeResolver: Resolver<
+const entityType: Resolver<
   Omit<
     GQLEntityType["entityType"],
     | EntityTypeTypeFields
@@ -57,30 +36,30 @@ const entityTypeResolver: Resolver<
   if (gqlEntityType.entityType) {
     return gqlEntityType.entityType;
   }
-  return getEntityTypeById(dataSources, gqlEntityType.entityId);
+  return getEntityTypeType(dataSources).then((type) => type.toGQLEntityType());
 };
 
 /**
  * Get the entityTypeId of an EntityType, i.e. the entityId of the "EntityType" EntityType.
  */
-const entityTypeIdResolver: Resolver<
+const entityTypeId: Resolver<
   GQLEntityType["entityTypeId"],
   EntityTypeMaybeTypeFields,
   GraphQLContext
-> = async (gqlEntityType, __, _) => {
+> = async (gqlEntityType, __, { dataSources }) => {
   if (gqlEntityType.entityType) {
     return gqlEntityType.entityType.entityId;
   }
   if (gqlEntityType.entityTypeId) {
     return gqlEntityType.entityTypeId;
   }
-  return gqlEntityType.entityId;
+  return getEntityTypeType(dataSources).then((type) => type.entityId);
 };
 
 /**
  * Get the entityTypeName of an EntityType, i.e. the name of the "EntityType" EntityType.
  */
-const entityTypeNameResolver: Resolver<
+const entityTypeName: Resolver<
   GQLEntityType["entityTypeName"],
   EntityTypeMaybeTypeFields,
   GraphQLContext
@@ -99,7 +78,7 @@ const entityTypeNameResolver: Resolver<
 /**
  * Get the entityTypeVersionId of an EntityType, i.e. the entityVersionId of the "EntityType" EntityType.
  */
-const entityTypeVersionIdResolver: Resolver<
+const entityTypeVersionId: Resolver<
   GQLEntityType["entityTypeVersionId"],
   EntityTypeMaybeTypeFields,
   GraphQLContext
@@ -114,41 +93,8 @@ const entityTypeVersionIdResolver: Resolver<
 };
 
 export const entityTypeTypeFields = {
-  entityType: entityTypeResolver,
-  entityTypeId: entityTypeIdResolver,
-  entityTypeName: entityTypeNameResolver,
-  entityTypeVersionId: entityTypeVersionIdResolver,
-};
-
-const entityTypeChildrenResolver: Resolver<
-  Promise<UnresolvedGQLEntityType[]>,
-  GQLEntityType,
-  GraphQLContext
-> = async (params, _, { dataSources: { db } }) => {
-  const { accountId, entityId: entityTypeId } = params;
-  const schema$ID = generateSchema$id(accountId, entityTypeId);
-  const schemaRef = schema$idRef(schema$ID);
-
-  const entityTypes = await EntityType.getEntityTypeChildren(db, { schemaRef });
-
-  return entityTypes.map((entityType) => entityType.toGQLEntityType());
-};
-
-const entityTypeParentsResolver: Resolver<
-  Promise<UnresolvedGQLEntityType[]>,
-  EntityTypeMaybeTypeFields,
-  GraphQLContext
-> = async (params, _, { dataSources }) => {
-  const { entityId: entityTypeId } = params;
-
-  const entityTypes = await EntityType.getEntityTypeParents(dataSources.db, {
-    entityTypeId,
-  });
-
-  return entityTypes.map((ent) => ent.toGQLEntityType());
-};
-
-export const entityTypeInheritance = {
-  entityTypeChildren: entityTypeChildrenResolver,
-  entityTypeParents: entityTypeParentsResolver,
+  entityType,
+  entityTypeId,
+  entityTypeName,
+  entityTypeVersionId,
 };
